backend/requirements.txt
Flask
Flask-Cors
google-generativeai
mysql-connector-python
werkzeug
requests
beautifulsoup4